♻️ refactor: Articles 커서 로직 Service 영역으로 이동 #87
