+++ src/main/java/com/example/umc7th/domain/article/service/query/ArticleQueryService.java
@@ -1,9 +1,9 @@
 package com.example.umc7th.domain.article.service.query;
 
+import com.example.umc7th.domain.article.controller.ArticleSearchCond;
+import com.example.umc7th.domain.article.dto.ArticleResponseDTO;
 import com.example.umc7th.domain.article.entity.Article;
-import org.springframework.data.domain.Slice;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 public interface ArticleQueryService {
@@ -12,7 +12,5 @@ public interface ArticleQueryService {
     List<Article> getArticles(String keyword);
     boolean hasComments(Long id);
 
-    Slice<Article> getArticlesOrderById(Long id, int size);
-    Slice<Article> getArticlesOrderByCreatedAt(LocalDateTime cursor, int size);
-    Slice<Article> getArticlesOrderByLikeNum(String cursor, int size);
+    ArticleResponseDTO.ArticleCursorPreviewListDTO getArticlesOrderBy(ArticleSearchCond sortCond, String cursor, int size);
 }
